requirements.txt
aiogram~=3.0.0
requests~=2.31.0
psutil~=5.9.5
Pillow~=10.0.0
opencv-python~=4.8.0.76